[4] renamed user class file to user_account
