+++ user_account/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
 
-from user_account.models.user import UserAccount
+from user_account.models.user_account import UserAccount
 
 admin.site.register(UserAccount)
