+++ user_account/api/urls.py
@@ -1,10 +1,11 @@
 from django.urls import path
 
-from user_account.api.views import UserCreateView, UserLoginView
+from user_account.api.views import UserCreateView, UserDetailView, UserLoginView
 
 app_name = "user_account"
 
 urlpatterns = [
     path("register", UserCreateView.as_view(), name="register"),
     path("login", UserLoginView.as_view(), name="login"),
+    path("user/<int:pk>", UserDetailView.as_view(), name="user_detail"),
 ]
